Fix critical blocker: Block header uses correct field types (u32 + padding)
PROBLEM: Block headers were using wrong field types
- Previously: wrote i64 (8 bytes) for len and count
- Spec requires: u32 (4 bytes) + explicit zero-padding

SOLUTION: Updated block header format to match Blender specification
- Changed write_block_header() signature: i64 → u32 for both data_len and count
- Added explicit zero padding at offsets 20-23 and 28-31
- Block header now correctly: 32 bytes total = code[4] + sdna_idx[u32] + old_ptr[u64] +
  data_len[u32] + zero[u32] + count[u32] + zero[u32]

IMPACT: This fixes all .blend files generated by starbreaker
- All exported block headers now conform to Blender's binary format specification
- Ensures compatibility with Blender 5.1.x and later versions

CHANGED:
- crates/starbreaker-blend/src/lib.rs:
  * write_block_header(): u32 parameters + explicit zero padding
  * write_block(): u32 count parameter + u32 data_len cast
- crates/starbreaker-3d/src/pipeline/blend_assembly.rs:
  * Updated num_attrs cast: i64 → u32
  * Updated totvert cast: i64 → u32

TESTING: starbreaker-blend crate compiles successfully
- Function signature changes propagate correctly to callers
- All write_block() calls updated with proper u32 casts

+++ crates/starbreaker-3d/src/pipeline/blend_assembly.rs
@@ -531,7 +531,7 @@ fn mesh_to_blend(
     write_block(&mut out, b"DATA", 0, mesh_mat_ptr, 1, &mesh_mat_array);
 
     // Attribute descriptor block (all Attribute structs concatenated)
-    write_block(&mut out, b"DATA", SDNA_IDX_ATTRIBUTE, attrs_ptr, num_attrs as i64, &attr_blob);
+    write_block(&mut out, b"DATA", SDNA_IDX_ATTRIBUTE, attrs_ptr, num_attrs as u32, &attr_blob);
 
     // Attribute name strings
     write_block(&mut out, b"DATA", 0, name_pos_ptr, 1, b"position\0");
@@ -632,7 +632,7 @@ fn mesh_to_blend(
                 
                 // Write MDeformVert data array itself
                 let mdv_raw_data = build_mdeformvert_array(&mdeformvert_data);
-                write_block(&mut out, b"DATA", SDNA_IDX_MDEFORMVERT, mdv_data_ptr, totvert as i64, &mdv_raw_data);
+                write_block(&mut out, b"DATA", SDNA_IDX_MDEFORMVERT, mdv_data_ptr, totvert as u32, &mdv_raw_data);
             }
             
             // Write CustomDataLayer
